Release v1.0.2

+++ Projects/B-U585I-IOT02A/Applications/FileX/Fx_NoR_Write_Read_File/FileX/App/app_filex.c
@@ -149,7 +149,7 @@ void fx_thread_entry(ULONG thread_input)
   }
 
   /* Print the absolute size of the NOR chip*/
-  printf("Total NOR Flash Chip size is: %lu bytes.\n", ospi_info.FlashSize);
+  printf("Total NOR Flash Chip size is: %lu bytes.\n", (unsigned long)ospi_info.FlashSize);
 
   /* Format the NOR flash as FAT */
   status =  fx_media_format(&nor_flash_disk,
