+++ Projects/STM32U575I-EV/Applications/FileX/Fx_Dual_Instance/FileX/App/app_filex.c
@@ -364,7 +364,7 @@ VOID fx_thread_two_entry(ULONG thread_input)
   }
 
   /* Print the absolute size of the NOR chip*/
-  printf("Total NOR Flash Chip size is: %lu bytes.\n", ospi_info.FlashSize);
+  printf("Total NOR Flash Chip size is: %lu bytes.\n", (unsigned long)ospi_info.FlashSize);
 
   /* Format the NOR flash as FAT */
   status =  fx_media_format(&nor_flash_disk,
